Reworked internal implementation of AsyncWorker to use Threads instead of TPL.
Reason: any exception occured inside Task are silently catched by it's own implementation and then thrown in Task.Wait() as an AggregateException. Because user commands handling initiated from the AsyncWorker this prevents user from know if (and when) the exception was thrown in his code. This behavior adds a lot of confusion for final developer and for code debugging.