Match type of arguments to min() and max()
max() and min() were changed from macros to templates in ESP8266 core for Arduino 2.4.0. This change causes calls to these functions with arguments of different types to cause a compilation error "error: no matching function for call to". This commit changes the arguments to these functions to the same type to fix compatibility with ESP8266.